requirements.txt
streamlit
biopython
pandas
matplotlib
fpdf
seaborn
